Moves checklist functions from card to cardModal component
