+++ src/Component/Card/Card.js
@@ -11,44 +11,6 @@ export default class Card extends React.Component {
         isModalOpen: false
     }
     
-    addCheckListItem = (itemToAdd) => {
-        if (itemToAdd) {
-            const newCheckList = [...this.props.card.checkListItems, itemToAdd];
-            const editedCard = {
-                ...this.props.card,
-                checkListItems : newCheckList
-            }
-            this.props.editCard(this.props.cardId, editedCard);
-        }
-    }
-
-    onChangeCheckListItem = itemClicked => {
-        this.props.onChangeCheckListItem(this.props.cardId, itemClicked);
-    }
-
-    onDeleteCheckListItem = index => {
-        const editedCheckList = this.props.card.checkListItems.filter((item, currIndex) => currIndex !== index);
-        const editedCard = {
-          ...this.props.card,
-          checkListItems : editedCheckList
-          }
-        this.props.editCard(this.props.cardId, editedCard);
-        console.log(editedCheckList); 
-    }
-
-  onToggleCheckBox = index => {
-      const toggledCheckListItem = this.props.card.checkListItems[index];
-      const editedCheckListItem = {item: toggledCheckListItem.item, complete: !toggledCheckListItem.complete};
-      const editedCheckList = this.props.card.checkListItems.map((item, currIndex) =>{
-          return (currIndex === index ? editedCheckListItem : item);
-      });
-      const editedCard = {
-        ...this.props.card,
-        checkListItems : editedCheckList
-        }
-      this.props.editCard(this.props.cardId, editedCard);
-      console.log(editedCheckList);
-    }
 
     toggleModal = () => {
         const { isModalOpen } = this.state
@@ -83,7 +45,12 @@ export default class Card extends React.Component {
 
                     )}
                 </Draggable>
-                <CardModal card={card} toggleModal={this.toggleModal} isModalOpen={isModalOpen} editCard={editCard}/>
+                <CardModal 
+                    card={card}
+                    cardId={cardId}
+                    toggleModal={this.toggleModal} 
+                    isModalOpen={isModalOpen} 
+                    editCard={editCard}/>
             </Fragment>
         )
     }
